mcp-server/requirements.txt
mcp>=1.0.0
mysql-connector-python>=8.0.33

